separate card back view into own file
