+++ CosmoPic/Views/TodayViews/TodayPhotoView.swift
@@ -63,26 +63,3 @@ struct TodayPhotoView: View {
   }
   .environmentObject(FavoritesViewModel())
 }
-
-struct TodayPhotoCardBackView: View {
-  let photoExplanation: String
-  let height: CGFloat
-  let width: CGFloat
-  let onInfoButtonTapped: () -> Void
-
-  var body: some View {
-    ScrollView {
-      Text(photoExplanation)
-    }
-    .padding()
-    .frame(width: width, height: height)
-    .background(.ultraThinMaterial)
-    .cornerRadius(20)
-    .clipped()
-    .onTapGesture {
-      withAnimation {
-        onInfoButtonTapped()
-      }
-    }
-  }
-}
